fix(login): confine state to dispatcher
Windows ASIO scheduling exposed output-buffer and Lua access from worker threads during concurrent login. Dispatch packet and game callbacks, and keep database storage hydration free of Lua events.

Advertise Astra features only after protocol ownership and before initial map serialization. Harden creature lifetime and state access, add regression coverage, and provide a server-side login/storage stress command.

Validated with MSVC Release, ASan, Valgrind Memcheck, Callgrind, cppcheck, focused clang-tidy, and the live 200k-operation stress run.

+++ src/events.h
@@ -89,8 +89,8 @@ public:
 	ReturnValue eventCreatureOnTargetCombat(Creature* creature, Creature* target);
 	void eventCreatureOnHear(Creature* creature, Creature* speaker, std::string_view words, SpeakClasses type);
 	void eventCreatureOnChangeZone(Creature* creature, ZoneType_t fromZone, ZoneType_t toZone);
-	void eventCreatureOnUpdateStorage(Creature* creature, const uint32_t key, const std::optional<int32_t> value,
-	                                  const std::optional<int32_t> oldValue, bool isSpawn);
+	void eventCreatureOnUpdateStorage(Creature* creature, const uint32_t key, const std::optional<int64_t> value,
+	                                  const std::optional<int64_t> oldValue, bool isSpawn);
 
 	// Party
 	bool eventPartyOnJoin(Party* party, Player* player);
